Correct for 14-bit int wraparound

+++ ass3.2_with_control_loop/src/control_loop/include/control_loop.hpp
@@ -28,8 +28,8 @@ private:
     double u[4];
     double y[4];
 
-    uint8_t prevChannel1;
-    uint8_t prevChannel2;
+    int prevChannel1;
+    int prevChannel2;
 protected:
     //Functions
     int initialising() override;
